refactor: rename common components subdir to shared for clarity
