+++ frontend/src/components/layout/Header.tsx
@@ -7,7 +7,7 @@ import { useRouter } from 'next/navigation'
 import { User as UserIcon, Star as StarIcon } from 'lucide-react'
 import { useAuth } from '@/context/AuthContext'
 import SearchBar from '@/components/search/SearchBar'
-import HighlightedText from '@/components/common/HighlightedText'
+import HighlightedText from '@/components/shared/HighlightedText'
 import Dropdown from '@/components/layout/Dropdown'
 
 const Header = ({
