+++ frontend/src/components/search/SearchInner.tsx
@@ -9,8 +9,8 @@ import type { Media } from '@/lib/media/types'
 import type { SearchAPIResponse } from '@/lib/search/types'
 import { fetchSearchResults } from '@/lib/search/api'
 import MediaCard from '@/components/media/MediaCard'
-import PageNavigator from '@/components/common/PageNavigator'
-import LoadingSpinner from '@/components/common/LoadingSpinner'
+import PageNavigator from '@/components/shared/PageNavigator'
+import LoadingSpinner from '@/components/shared/LoadingSpinner'
 
 const SearchInner = () => {
   const params = useSearchParams()
